Added DropDown Lists and Checkboxes

+++ web.py
@@ -1,6 +1,5 @@
 from flask import Flask, render_template, request, send_file
 from datetime import date, datetime, timedelta
-from nwshelf import NWSHELF_Deenish_profile
 from CA import oceancolour
 from SST import SST_request
 from pytz import timezone
@@ -10,6 +9,8 @@ from PIL import Image
 from math import atan2, pi
 import Plot
 import os
+
+units = {'Temperature': '°C', 'Salinity': '', 'pH': '', 'RFU': '', 'Oxygen Saturation': '%'}
 
 def web(DBO):
     # Start application
@@ -23,18 +24,34 @@ def web(DBO):
         while t0 < t1:
             times.append(t0)
             t0 += timedelta(days=1)
-        times.append(t0)            
+        times += [t0, '']
+
+        # Get list of layers available in the Northwest Shelf model
+        layers = DBO.NWSHELF[0]
+        
+        variables = ['Temperature', 'Salinity', 'pH', 'Oxygen Saturation', 'RFU', '']
         
         if request.method == 'POST':
             
-            if 'Display' in request.form:                                
-                sub, t0, t1 = subs(DBO, request)  
-                # Remove older images
+            if 'Display' in request.form:   
+                
+                ''' Remove older images '''
                 images = glob('static/Deenish*')
                 for file in images:
                     os.remove(file) 
+                images = []
+                           
+                images += subs(DBO, request)  
+                
+                ''' Get user's temperature layer selection '''
+                z = request.form.getlist('layers')
+                
+                '''' Plot temperature profile from MetOffice's Northwest Shelf model '''
+                if z: profile = subset_nwshelf(DBO, z); images += profile
+                
+                ''' Retrieve date selection for surface currents '''  
                 # Date for surface currents information
-                datestring = ''
+                arrows = []; datestring = ''
                 if request.form.get('times'):
                     datestring = request.form.get('times')
                     # Get date
@@ -42,38 +59,34 @@ def web(DBO):
                     # Get currents for the selected date
                     u, v, t = surface_currents_request(DBO, fecha)                      
                     # Plot surface currents
-                    arrows = Plot.Plot_Arrows(u, v, t)                                        
-                    # Resize
-                    A = Image.open(f'static/{arrows}').resize((1520, 1520), Image.ANTIALIAS)
-                    A.save(f'static/{arrows}', quality=95)
-                else:
-                    arrows = 'blank.png'                              
-                # Read temperature profile from MetOffice's Northwest Shelf model
-                time, temp = NWSHELF_Deenish_profile(t0, t1)                               
-                # Plot time series
-                name = Plot.Plot_Deenish_user_selection(sub, DBO, time, temp) 
-                # Resize
-                I = Image.open(f'static/{name}').resize((1520, 2267), Image.ANTIALIAS)
-                I.save(f'static/{name}', quality=95)                
-                # Reload page with new figure
-                return render_template('index.html', times=times, graph=name,
-                    arrows=arrows, time0=t0.date(), time1=t1.date()-timedelta(days=1),
-                    time=datestring)                 
+                    arrows = Plot.Plot_Arrows(u, v, t)                                                           
+                images += arrows
+                                
+                ''' Reload page with new figures '''                
+                return render_template('index.html', times=times, graph=images,
+                   time=datestring, layers=layers, variables=variables)                 
                 
             elif 'Download' in request.form:
-                sub, _, _ = subs(DBO, request)
-                # Remove old CSV files
-                lista = glob('static/*.csv')
-                for file in lista:
-                    os.remove(file)
-                # Generate CSV file
-                f = to_csv(sub)                
-                # Download CSV file
-                return send_file(f, as_attachment=True)
+                sub, _, _ = get_requested_data(DBO, request)
+                user_variables = request.form.getlist('selection')
+                if sub:
+                    # Remove old CSV files
+                    lista = glob('static/*.csv')
+                    for file in lista:
+                        os.remove(file)
+                    # Generate CSV file
+                    f = to_csv(sub, user_variables)                
+                    # Download CSV file
+                    return send_file(f, as_attachment=True)
+                else:
+                    return render_template('index.html', times=times, graph=['blank.png'],
+                        arrows='blank.png', time0=times[-1], time1=times[-1], 
+                        time='', layers=layers, variables=variables) 
          
         else:
-            return render_template('index.html', times=times, graph='blank.png',
-                arrows='blank.png', time0=times[0], time1=times[-1], time='') 
+            return render_template('index.html', times=times, graph=['blank.png'],
+                arrows='blank.png', time0=times[-1], time1=times[-1], 
+                time='', layers=layers, variables=variables) 
         
     
     @app.route('/swirl', methods=['GET', 'POST'])
@@ -113,16 +126,82 @@ def web(DBO):
     
     app.run(debug=False)
     
+def subset_nwshelf(DBO, Z):
+    ''' Subset data from NORTHWEST SHELF model for given times and depth '''
+    
+    image = []
+    
+    depth, time, temp = DBO.NWSHELF
+    
+    _, t0, t1 = get_requested_data(DBO, request)
+        
+    if t0 and t1:
+        
+        # Find time indexes (start and end)
+        t0, t1 = np.argmin(abs(t0 - time)), np.argmin(abs(t1 - time))
+        # Subset time
+        time = time[t0 : t1]
+    
+        temperature = []
+        for z in Z:
+            z = int(z)
+            # Find depth index
+            z = np.argmin(abs(z - depth))
+            # Subset time and temperature    
+            temperature.append(temp[t0 : t1, z])
+           
+        
+        image += Plot.Plot_NWSHELF_Profile((time, Z, temperature))
+        
+            
+    return image
+
+def get_requested_data(DBO, request):
+    
+    t0, t1 = request.form.get('times0'), request.form.get('times1')
+    
+    sub = None
+    
+    if t0 and t1:
+        
+        # Make sure input times are within available time list
+        t0, t1 = fix_times(DBO, t0, t1)
+            
+        # Return subset
+        sub = subset(DBO.buoy, t0, t1)
+        
+    return sub, t0, t1
+
     
 def subs(DBO, request):
     ''' Get data buoy subset for the requested time period '''
     
-    # Retrieve request
-    t0, t1 = request.form.get('times0'), request.form.get('times1')
-    # Make sure input times are withing available time list
-    t0, t1 = fix_times(DBO, t0, t1)
-    # Return subset
-    return subset(DBO.buoy, t0, t1), t0, t1
+    images = []
+        
+    ''' Get user's variable selection '''
+    user_variables = request.form.getlist('selection')
+    
+    sub, t0, t1 = get_requested_data(DBO, request)
+    
+    ''' Get user's YY-Plots selection '''
+    A1, A2 = request.form.get('A1'), request.form.get('A2')
+    B1, B2 = request.form.get('B1'), request.form.get('B2')
+    C1, C2 = request.form.get('C1'), request.form.get('C2')
+    YY = {'A1': A1, 'A2': A2, 'B1': B1, 'B2': B2, 'C1': C1, 'C2': C2}
+    
+    if t0 and t1:
+        
+        ''' Plot single time series of user's selection '''
+        for variable in user_variables:
+            name = Plot.Plot_Deenish_user_selection(sub, DBO, variable) 
+            images += name
+            
+                    
+        names = Plot.Plot_Deenish_YY(sub, YY)
+        
+        images += names
+        
+    return images
     
 def fix_times(DBO, t0, t1):
     ''' Make sure that input times are within the available time list '''
@@ -170,39 +249,43 @@ def subset(buoy, t0, t1):
     # Find appropriate time indexes
     i0, i1 = np.argmin(abs(time - t0)), np.argmin(abs(time - t1)) + 1
     
-    sub = {}
-    for i in buoy.keys():
+    sub = {'time': buoy['time'][i0 : i1]}
+    for i in buoy.keys():        
         # Subset each parameter for the requested time period
         sub[i] = buoy[i][i0 : i1]
         
     return sub
 
 
-def to_csv(sub):
+def to_csv(sub, user_variables):
     ''' Write data subset into CSV file for download '''
     
     # Set output file name    
     f = 'static/Deenish-' + datetime.now().strftime('%Y%m%d%H%M%S') + '.csv'
     
-    with open(f, 'w') as csvfile:       
-        csvfile.write('Date\tTemperature (°C)\tSalinity\tpH\tRFU\tDissolved Oxygen Saturation (%)\tSpeed (cm/s)\tDirection\n')
+    # Initialize header
+    header = 'Date\t'
+    
+    with open(f, 'w') as csvfile:  
+        # Write header
+        for k in sub.keys():
+            if k != 'time' and k in user_variables:
+                # Append columns to header
+                header += ( k + ' ' + units[k] + '\t' )
+        # Write header
+        csvfile.write(header + '\n')
         for i, time in enumerate(sub['time']):
-            t = time.strftime('%d-%b-%Y %H:%M')
-            T = sub['temp'][i]
-            S = sub['salt'][i]
-            p = sub['pH'][i]
-            C = sub['chl'][i]
-            O = sub['DOX'][i]
-            u, v = sub['u'][i], sub['v'][i]
-            V = (u**2 + v**2)**.5
-            D = 90 - atan2(v, u) * 180 / pi
-            if D < 0: D += 360
-            
-            try:
-                l = '%s\t%.3f\t%.3f\t%.3f\t%.3f\t%.3f\t%.1f\t%.03d\n' % (t, T, S, p, C, O, V, D)
-            except ValueError:
-                l = '%s\t%.3f\t%.3f\t%.3f\t%.3f\t%.3f\t%.1f\t%f\n' % (t, T, S, p, C, O, V, D)
-            csvfile.write(l)
+            l = time.strftime('%d-%b-%Y %H:%M') + '\t'
+            for k in sub.keys():
+                if k != 'time' and k in user_variables:
+                    l += '%.3f\t' % sub[k][i]
+                
+                    # u, v = sub['u'][i], sub['v'][i]
+                    # V = (u**2 + v**2)**.5
+                    # D = 90 - atan2(v, u) * 180 / pi
+                    # if D < 0: D += 360
+                
+            csvfile.write(l + '\n')
         
     return f
